+++ CHANGELOG.md
@@ -11,11 +11,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   fnLs, added metadata option to return more file metadata (size, created, etc...)
 -   vars_files on a form.  Allowing to load multiple yaml files that can be attached to the form and will be available as contants.  It must be a dictionary and the keys should be unique to avoid conflicts with other fieldnames.
+-   minValue, maxValue, minLength, maxLength, minSize and maxSize are now allowing placeholder making them dynamic.  
 
 ### Fixed
 
 -   approval message placeholder
 -   small visual job issue
+-   OIDC login (kudos to theymademedothat)
 
 ### Changed